backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
pytest==7.4.3
black==23.11.0
flake8==6.1.0
